+++ CHANGELOG.md
@@ -3,6 +3,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased](https://github.com/h2oai/datatable/compare/HEAD...v0.2.2)
+#### Added
+- Method `df.tonumpy()` now has argument `stype` which will force conversion into
+  a numpy array of the specific stype.
+
 #### Changed
 - When writing "round" doubles/floats to CSV, they'll now always have trailing zero.
   For example, [0.0, 1.0, 1e23] now produce "0.0,1.0,1.0e+23" instead of "0,1,1e+23".